Making s3_to_redshift operator complete
